+++ gui/brassica-gui/ruleshighlighter.h
@@ -11,16 +11,20 @@ public:
     RulesHighlighter(QTextDocument *parent);
 
     void setCategories(QStringList categories, bool forceUpdate = false);
+    void setHighlights(QList<int> highlights);
 
 protected:
     void highlightBlock(const QString &text) override;
 
 private:
+    QList<int> highlights;
+
+    QTextCharFormat categoryFormat;
+    QTextCharFormat highlightFormat;
+
     // NB. these two shouldn't get out of sync!
     QStringList m_categories;
     QRegularExpression categoriesPattern;
-
-    QTextCharFormat categoryFormat;
 
     // and neither should these
     QVector<QTextCharFormat> formats;